chore: gate pytest-httpserver fork test on python>=3.10
The pytest-httpserver dev dependency requires python>=3.10. Move the
fork test (the only consumer) into its own file so the four
pre-existing polling-manager tests still run on 3.9, and skip the new
file there via importorskip. Add a mypy override so 3.9 type-checks
do not error on the missing module.
