+++ pyproject.toml
@@ -35,6 +35,12 @@ pytest-httpserver = {version = "^1.1.0", python = ">=3.10,<4"}
 [tool.mypy]
 exclude = ["example/*"]
 
+[[tool.mypy.overrides]]
+# pytest-httpserver requires python>=3.10, so it is not installed in
+# the 3.9 dev environment and mypy cannot find its stubs there.
+module = "pytest_httpserver"
+ignore_missing_imports = true
+
 [tool.black]
 target-version = ["py39"]
